+++ src/main/java/com/server/service/EmployeeService.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface EmployeeService {
 	public long createEmployee(Employee employee);
     public Employee updateEmployee(Employee employee);
-    public void deleteEmployee(long id);
+    public void deleteEmployee(int id);
     public List<Employee> getAllEmployees();
-    public Employee getEmployee(long id);	
+    public Employee getEmployee(int id);
 	public List<Employee> getAllEmployees(String employeeName);
 }
